Update DC sweep tests for version 0.0.21
There was a bug in netlist generation in qucs-s versions prior to
0.0.21.  It was fixed in 0.0.21.

The 01-SuperSimpleDC and 02-DiodeClipper examples had previously been
set up to work around this bug.  I am removing the workaround, and
changing the READMEs to reflect that the bug has been fixed, and what
workaround is necessary if using an older version of qucs-s.