Add an example of S-parameter extraction
Per ra3xdh/qucs_s#79, qucs-s is getting support for Xyce S parameter
extraction.  This example demonstrates that usage.

At this time, it will only work in qucs-s from the "current" branch,
as the capability is not in any released versions.  It is slated for
support in qucs-s 0.0.24.
ra3xdh/qucs_s#79