Update example to use new expression library
Since Xyce 7.2, the expression library has begun to support complex
arithmetic, and this example, which served to show how to compute K
using explicit real arithmetic, is no longer the best way to work.

I have modified the example to make it compute K exactly as ngspice
does, using db(v(out)/v(in)).

This modification requires that a recent version of Xyce be used.