Update examples for qucs-s 0.0.24
A few of my example schematics threw warnings upon being loaded into
qucs-s 0.0.24.  These warnings were harmless, but reflected the
presence in the .sch file of artifacts of an old run on 0.0.21.  As
soon as they were loaded into 0.0.24 these artifacts were removed and
replaced with new information, but required that the changes be saved.

This commit is the result of loading all of them one at a time, saving
any that were changed, and rerunning them to make sure they still
worked.

I am no longer able to install Octave on my systems, and cannot test
out the netlists in 08-OctaveScripting.

Note that 08-OctaveScripting, while it is based on examples in the
qucs-s manual, fixes some mistakes in the octave modules in that
manual that rendered the examples broken.