Mark 08-OctaveScripting as broken
Since I created them, both Xyce and Qucs-s have changed enough that
the Octave scripts in this directory are now incorrectly attempting to
access HB output information from the .dat.xyce file.

I have limited access to octave and am no longer able to maintain this
test.  But rather than just delete it, I'm adding information to the
README to mark it broken and ask anyone who wishes to do so to help me
fix it.