Fix VCC setting in two examples
I went back and reviewed the old Xyce app note for the common emitter
circuits and the old ARRL "Hands On Radio" articles from which they
were derived, and found that my Xyce_qucs-s_Examples for
03_CommonEmitter_flat and 04_CommonEmitter_subcircuited both have vcc
set wrong, and therefore plot input and output voltages that are far
too small.  The circuit still ran, but the results were silly.

The correct VCC value was supposed to be 12V, and I had neglected to
change the default value of the DC source for VCC when I placed the
item in the Qucs-s schematic.

Now the plots produced agree with those in the original article I had
used to develop these examples in the first place.